--- app/src/main/java/android/example/com/split/AddExpenseActivity.java
@@ -1,13 +1,0 @@
-package android.example.com.split;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class AddExpenseActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_expense);
-    }
-}
