--- app/src/main/java/android/example/com/split/LedgerActivity.java
@@ -1,13 +1,0 @@
-package android.example.com.split;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class LedgerActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_ledger);
-    }
-}
