+++ app/src/main/java/android/example/com/split/data/entity/User.java
@@ -11,6 +11,7 @@ import java.util.List;
  */
 @IgnoreExtraProperties
 public class User {
+
     // firebase authentiction id
     private String authId;
 
@@ -37,10 +38,6 @@ public class User {
 
     private List<String> contactList;
 
-    public List<String> getContactList() {
-        return contactList;
-    }
-
     /**
      * User
      */
@@ -48,6 +45,7 @@ public class User {
     public User() {
 
     }
+
     //
     public User(String firstName, String lastName, String email, String phoneNumber) {
 
@@ -68,6 +66,10 @@ public class User {
         contactList = new ArrayList<String>();
     }
 
+    public List<String> getContactList() {
+        return contactList;
+    }
+
     @Exclude
     public String getId() {
         return id;
@@ -86,13 +88,13 @@ public class User {
 
     }
 
-
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
 
     public String getLastName() {
         return lastName;
     }
-
-
 
     public String getEmail() {
         return email;
@@ -109,12 +111,9 @@ public class User {
     public void setPhoneNumber(String phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
-    // add new member to my contact list
-    public void addToContactList(String user){
-        contactList.add(user);
-    }
 
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
+    // add new member to my contact list
+    public void addToContactList(String user) {
+        contactList.add(user);
     }
 }
