+++ app/src/main/java/android/example/com/split/ui/home/HomeActivity.java
@@ -1,10 +1,11 @@
-package android.example.com.split;
+package android.example.com.split.ui.home;
 
 import android.app.ActionBar;
 import android.app.AlertDialog;
 import android.content.Intent;
+import android.example.com.split.R;
+import android.example.com.split.ui.FullscreenActivity;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.NavigationView;
 import android.support.design.widget.TabLayout;
@@ -19,10 +20,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
-import com.firebase.ui.auth.AuthUI;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 
 public class HomeActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
@@ -74,7 +71,7 @@ public class HomeActivity extends AppCompatActivity implements NavigationView.On
 
                     /*Fragment f = getSupportFragmentManager().findFragmentById(R.id.tabLayout);
 
-                    if (f instanceof GroupsFragment){
+                    if (f instanceof GroupsTabFragment){
                         Toast.makeText(getBaseContext(), "Groups", Toast.LENGTH_LONG).show();
                     }
 
@@ -95,7 +92,7 @@ public class HomeActivity extends AppCompatActivity implements NavigationView.On
             navigationView.setNavigationItemSelectedListener(this);
 
             ViewPager viewPager = (ViewPager) findViewById(R.id.viewpager);
-            CategoryAdapter adapter = new CategoryAdapter(this, getSupportFragmentManager());
+            HomeTabsAdapter adapter = new HomeTabsAdapter(this, getSupportFragmentManager());
             viewPager.setAdapter(adapter);
 
             FloatingActionButton floatingActionButton = findViewById(R.id.fab);
@@ -281,7 +278,7 @@ public class HomeActivity extends AppCompatActivity implements NavigationView.On
 
     private void createContactPopupDialog() {
         dialogBuilder = new AlertDialog.Builder(this);
-        View view = getLayoutInflater().inflate(R.layout.add_contact_popup, null);
+        View view = getLayoutInflater().inflate(R.layout.dialog_add_contact, null);
         contactItem = (EditText) findViewById(R.id.contactItem);
         saveButton = (Button) findViewById(R.id.saveContactButton);
 
@@ -293,7 +290,7 @@ public class HomeActivity extends AppCompatActivity implements NavigationView.On
 
     private void createGroupPopupDialog() {
         dialogBuilder = new AlertDialog.Builder(this);
-        View view = getLayoutInflater().inflate(R.layout.add_group_popup, null);
+        View view = getLayoutInflater().inflate(R.layout.dialog_add_group, null);
         contactItem = (EditText) findViewById(R.id.groupItem);
         saveButton = (Button) findViewById(R.id.saveGroupButton);
