+++ app/src/main/java/android/example/com/split/ui/home/groups/GroupsRecyclerAdapter.java
@@ -1,13 +1,14 @@
-package android.example.com.split;
+package android.example.com.split.ui.home.groups;
 
 import android.content.Intent;
+import android.example.com.split.R;
 import android.example.com.split.data.entity.Group;
+import android.example.com.split.ui.home.groups.group.GroupDetailActivity;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,16 +19,49 @@ public class GroupsRecyclerAdapter extends RecyclerView.Adapter<GroupsRecyclerAd
 
     private List<android.example.com.split.data.entity.Group> mDataset;
 
+    // Create the adapter with a dataset
+    public GroupsRecyclerAdapter(List<Group> myDataset) {
+        mDataset = myDataset;
+    }
+
+    // Create new views (invoked by the layout manager)
+    @Override
+    public GroupRowViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+
+        View v = (View) LayoutInflater.from(parent.getContext()).inflate(R.layout.item_group, parent, false);
+        GroupRowViewHolder vh = new GroupRowViewHolder(v, this);
+        return vh;
+    }
+
+    // Replace the contents of a view (invoked by the layout manager)
+    @Override
+    public void onBindViewHolder(GroupRowViewHolder holder, int position) {
+        // - get element from your dataset at this position
+        // - replace the contents of the view with that element
+        android.example.com.split.data.entity.Group group = mDataset.get(position);
+        holder.mTextView.setText(group.getName());
+        holder.expenseTextView.setText("" + group.getExpenses().get(0).getPaymentAmount());
+    }
+
+    // Return the size of your dataset (invoked by the layout manager)
+    @Override
+    public int getItemCount() {
+        if (mDataset == null) {
+            mDataset = new ArrayList<>();
+        }
+        return mDataset.size();
+    }
+
     // Provides reference to the views for each data item
     // When create more complex group view, it should be removed in a separate java file
     class GroupRowViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
+        final GroupsRecyclerAdapter mAdapter;
         // Each group data item is just a String presented as a textView in this case
         public TextView mTextView;
         public TextView expenseTextView;
-        final GroupsRecyclerAdapter mAdapter;
 
-        // Initializes the ViewHolder TextView from the group_row_view XML resource
+        // Initializes the ViewHolder TextView from the item_group XML resource
         public GroupRowViewHolder(View v, GroupsRecyclerAdapter adapter) {
             super(v);
             mTextView = (TextView) v.findViewById(R.id.groupRowView);
@@ -50,39 +84,6 @@ public class GroupsRecyclerAdapter extends RecyclerView.Adapter<GroupsRecyclerAd
             mAdapter.notifyDataSetChanged();
             v.getContext().startActivity(intent);
         }
-    }
-
-    // Create the adapter with a dataset
-    public GroupsRecyclerAdapter(List<Group> myDataset) {
-        mDataset = myDataset;
-    }
-
-    // Create new views (invoked by the layout manager)
-    @Override
-    public GroupRowViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-
-        View v = (View) LayoutInflater.from(parent.getContext()).inflate(R.layout.group_row_view, parent, false);
-        GroupRowViewHolder vh = new GroupRowViewHolder(v, this);
-        return vh;
-    }
-
-    // Replace the contents of a view (invoked by the layout manager)
-    @Override
-    public void onBindViewHolder(GroupRowViewHolder holder, int position) {
-        // - get element from your dataset at this position
-        // - replace the contents of the view with that element
-        android.example.com.split.data.entity.Group group = mDataset.get(position);
-        holder.mTextView.setText(group.getName());
-        holder.expenseTextView.setText("" + group.getExpenses().get(0).getPaymentAmount());
-    }
-
-    // Return the size of your dataset (invoked by the layout manager)
-    @Override
-    public int getItemCount() {
-        if(mDataset==null){
-            mDataset = new ArrayList<>();
-        }
-        return mDataset.size();
     }
 }
