--- app/src/main/res/layout/activity_add_member.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    tools:context=".AddMemberActivity"
-    tools:layout_editor_absoluteY="81dp"
-    android:gravity="center">
-
-    <RelativeLayout
-        android:id="@+id/relative_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ImageView
-            android:id="@+id/back"
-            android:layout_alignParentStart="true"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="back"
-            app:srcCompat="@drawable/baseline_arrow_back_black_24dp" />
-
-
-        <Button
-            android:id="@+id/done_addmember_button"
-            android:layout_alignParentEnd="true"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="225dp"
-            android:onClick="done"
-            android:paddingStart="20dp"
-            android:text="@string/button_lable_done"
-            app:layout_constraintEnd_toEndOf="parent"
-            tools:layout_editor_absoluteY="0dp" />
-
-
-    </RelativeLayout>
-
-        <TextView
-            android:id="@+id/show_groupname"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_above="@+id/linear_layout"
-            android:text="@string/groupname"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            tools:layout_editor_absoluteX="144dp"
-            tools:layout_editor_absoluteY="68dp"
-            android:paddingBottom="20dp"/>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <EditText
-            android:id="@+id/addNameEditText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:ems="10"
-            android:hint="@string/hint_add_member"
-            android:inputType="textPersonName" />
-
-        <Button
-            android:id="@+id/add_member_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:text="@string/button_lable_addmember" />
-    </RelativeLayout>
-
-    <ListView
-        android:id="@+id/listv"
-        android:layout_width="match_parent"
-        android:layout_height="429dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginBottom="30dp"
-        android:paddingLeft="20dp"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="159dp">
-
-    </ListView>
-
-
-</LinearLayout>
