--- app/src/main/res/layout/activity_ledger.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    tools:context=".LedgerActivity"
-    tools:layout_editor_absoluteY="81dp"
-    android:gravity="center">
-
-
-    <TextView
-        android:id="@+id/show_groupname"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/groupname"
-        android:textSize="18sp"
-        android:textStyle="bold"
-        tools:layout_editor_absoluteX="144dp"
-        tools:layout_editor_absoluteY="68dp"
-        android:paddingBottom="20dp"/>
-
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ListView
-            android:id="@+id/ledgerListView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@id/addExpense"/>
-
-        <ImageButton
-            android:id="@+id/addMemberFromLedger"
-            android:layout_width="wrap_content"
-            android:layout_below="@id/ledgerListView"
-            android:layout_height="wrap_content"
-            android:contentDescription="Add a member"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentBottom="true"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:src="@drawable/baseline_person_add_black_24dp"
-            />
-
-        <ImageButton
-            android:id="@+id/addExpense"
-            android:layout_width="wrap_content"
-            android:contentDescription="Add an expense"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentBottom="true"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:src="@drawable/baseline_shopping_cart_black_24dp"
-            />
-
-
-    </RelativeLayout>
-
-</LinearLayout>
