+++ app/src/main/res/layout/dialog_add_contact.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:id="@+id/layout_id"
+    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
