+++ app/src/main/res/layout/item_group.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="@dimen/list_item_height"
     android:layout_marginLeft="@dimen/margin_medium"
@@ -10,8 +8,7 @@
     android:gravity="center_vertical">
 
     <!-- A CardView that contains a TextView -->
-    <android.support.v7.widget.CardView
-        xmlns:card_view="http://schemas.android.com/apk/res-auto"
+    <android.support.v7.widget.CardView xmlns:card_view="http://schemas.android.com/apk/res-auto"
         android:id="@+id/card_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -31,17 +28,26 @@
                 android:layout_height="match_parent"
                 android:layout_weight="0"
                 android:scaleType="centerInside"
-                app:srcCompat="@drawable/ic_menu_share" />
+                app:srcCompat="@drawable/ic_group_black_24dp" />
 
             <TextView
-                android:id="@+id/title"
+                android:id="@+id/groupRowView"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_weight="1"
-                android:paddingLeft="20dp"
                 android:text="@string/group_row_view"
                 android:textColor="@color/colorPrimaryDark"
-                android:textStyle="bold" />
+                android:textStyle="bold"
+                android:paddingLeft="20dp" />
+
+            <TextView
+                android:id="@+id/groupExpense"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="0"
+                android:textColor="@color/colorPrimaryDark"
+                android:textStyle="bold"
+                android:text="TextView" />
         </LinearLayout>
     </android.support.v7.widget.CardView>
